requirements.txt
pyyaml
tinydb
fire
matplotlib
flask
ujson